+++ src/game.h
@@ -632,9 +632,9 @@ class Game
 		void addCreatureHealth(const CreatureConstPtr& target);
 		void addCreatureHealth(const CreatureConstPtr& target, std::span<const CreaturePtr> spectators);
 		static void addCreatureHealth(const SpectatorVec& spectators, const CreatureConstPtr& target);
-		void addMagicEffect(const Position& pos, const uint8_t effect, std::span<const CreaturePtr> pre_cache);
-		void addMagicEffect(const Position& pos, uint8_t effect);
-		static void addMagicEffect(const SpectatorVec& spectators, const Position& pos, uint8_t effect);
+		void addMagicEffect(const Position& pos, const uint16_t effect, std::span<const CreaturePtr> pre_cache);
+		void addMagicEffect(const Position& pos, uint16_t effect);
+		static void addMagicEffect(const SpectatorVec& spectators, const Position& pos, uint16_t effect);
 		void addDistanceEffect(const Position& fromPos, const Position& toPos, uint8_t effect);
 		void addDistanceEffect(std::span<const CreaturePtr> spectators, const Position& fromPos, const Position& toPos, uint8_t effect);
 		static void addDistanceEffect(const SpectatorVec& spectators, const Position& fromPos, const Position& toPos, uint8_t effect);
